Changed Context-enabled widget to only connect to sibling dom nodes
Tooltips were not working in a repeated situation such as in a list
view, as the selector was only keying off of a CSS class.

In addition to the CSS class selector, the widget has also been updated
to only apply a tooltip to siblings of the widget itself. Therefore, a
tooltip widget and its target must be in the same container.